Integrate command maker into Luminol, instead of it being a separate application
