+++ Cargo.toml
@@ -145,7 +145,7 @@ opt-level = 3
 # https://bevy-cheatbook.github.io/pitfalls/performance.html#why-not-use---release
 
 [workspace]
-members = ["alox-48", "command-gen", "command-lib", "rmxp-types"]
+members = ["alox-48", "command-lib", "rmxp-types"]
 
 [patch.crates-io]
 poll-promise = { git = "https://github.com/EmbarkStudios/poll-promise", rev = "8c4d0d6e4d693cf54e9306950cb2f6d271c77080" }
