--- command-gen/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "command-gen"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-eframe = "0.21.3"
-
-command-lib = { version = "*", path = "../command-lib" }
-ron = "0.8.0"
-
-strum = { version = "0.24" }
-
-rand = "0.8.5"
